requirements.txt
extra_streamlit_components==0.1.60
matplotlib==3.7.1
numpy==1.24.3
Pillow==9.5.0
Pillow==10.0.0
streamlit==1.23.1
